# packages/api-client/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/audit-logs": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "list_audit_logs",
        "responses": {
          "200": {
            "description": "List of system audit logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_health",
        "responses": {
          "200": {
            "description": "System health status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "list_orders",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Keyword filter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Order status filter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination limit",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of filtered orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "handlers"
        ],
        "operationId": "create_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order unique ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "put": {
        "tags": [
          "handlers"
        ],
        "operationId": "update_order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "delete": {
        "tags": [
          "handlers"
        ],
        "operationId": "delete_order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditLog": {
        "type": "object",
        "required": [
          "id",
          "log_type",
          "user_name",
          "action",
          "status",
          "created_at"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "log_type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "required": [
          "client",
          "items",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "client": {
            "type": "string"
          },
          "items": {
            "type": "string"
          },
          "priority": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status",
          "version",
          "uptime_seconds",
          "database"
        ],
        "properties": {
          "database": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "uptime_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
      "Order": {
        "type": "object",
        "required": [
          "id",
          "client",
          "items",
          "amount",
          "status",
          "priority",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "client": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "items": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "UpdateOrderRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "client": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "enterprise-api",
      "description": "Enterprise Vibe-Coding Backend API"
    }
  ]
}